% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_gh_bbox_sum}
\alias{add_gh_bbox_sum}
\title{Determine total sum for each geohash with buffer of 200m}
\usage{
add_gh_bbox_sum(hash_decode, full, lon_nm, lat_nm, radius)
}
\arguments{
\item{hash_decode}{data.frame as output from `geohashTools::gh_decode()`}

\item{full}{data.frame with portfolio with lon, lat and sum insured}
}
\description{
Determine total sum for each geohash with buffer of 200m
}
\keyword{internal}
