% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_gh_nghbrs_sum}
\alias{add_gh_nghbrs_sum}
\title{Determine total insured sum of all neighbours (including hash itself)
df should contain a column called hash and a column with sum per hash}
\usage{
add_gh_nghbrs_sum(df, colname_sum)
}
\description{
Determine total insured sum of all neighbours (including hash itself)
df should contain a column called hash and a column with sum per hash
}
\keyword{internal}
