% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_around_coord.R
\name{block_around_coord}
\alias{block_around_coord}
\title{Title}
\usage{
block_around_coord(lon = lon, lat = lat, radius = 200)
}
\arguments{
\item{lon}{Longitude of point(s), in degrees.}

\item{lat}{Latitude of point(s), in degrees.}

\item{radius}{Major (equatorial) radius (default is meters) of the ellipsoid. The default value is for WGS84.}
}
\description{
Title
}
