% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_around_point.R
\name{block_around_point}
\alias{block_around_point}
\title{Coordinates of the four cardinal directions from center}
\usage{
block_around_point(lon_center, lat_center, radius = 200)
}
\arguments{
\item{lon_center}{Longitude of center point, in degrees.}

\item{lat_center}{Latitude of center point, in degrees.}

\item{radius}{Major (equatorial) radius (default is meters) of the ellipsoid. The default value is for WGS84.}
}
\value{
A data.frame consisting of the four coordinates (long,lat) obtained by walking the radius to the four cardinal directions.
}
\description{
Coordinates of the four cardinal directions from center
}
\examples{
block_around_point(lon_center = 6.520386, lat_center = 53.24007, radius = 200)
}
