% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cells_above_threshold}
\alias{cells_above_threshold}
\title{Identify the focal cells exceeding the threshold}
\usage{
cells_above_threshold(focal, threshold)
}
\arguments{
\item{focal}{focal as output from \code{terra::focal()}.}

\item{threshold}{lower (numeric) threshold boundary.}
}
\description{
Generate a data.frame containing the cell indices of the focal
cells surpassing the specified threshold. Additionally, include columns for
the coordinates (xy) corresponding to the center of each cell.
}
\author{
Martin Haringa
}
\keyword{internal}
