% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conc_per_cell_new}
\alias{conc_per_cell_new}
\title{Determine the concentrations within the highest focal cells for the current
iteration}
\usage{
conc_per_cell_new(high_foc, dff, value, size, points, db, radius)
}
\arguments{
\item{high_foc}{data.frame containing cell ids with the top n focal values
from the current iteration.}

\item{dff}{data.frame with all observations}

\item{value}{column name in `dff` to find concentrations for.}

\item{size}{size of cell in meters.}

\item{points}{number of points per `size`.}

\item{db}{data.frame containing previously saved highest concentrations.}

\item{radius}{radius of circle in meters.}
}
\description{
Determine the concentrations within the highest focal cells for
the current iteration.
}
\author{
Martin Haringa
}
\keyword{internal}
