% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-aliases.R
\name{concentration}
\alias{concentration}
\title{Deprecated alias for radius_sum()}
\usage{
concentration(
  sub,
  full,
  value,
  lon_sub = lon,
  lat_sub = lat,
  lon_full = lon,
  lat_full = lat,
  radius = 200,
  display_progress = TRUE,
  result_col = "radius_sum"
)
}
\arguments{
\item{sub}{Deprecated. Use \code{targets} in [radius_sum()] instead.}

\item{full}{Deprecated. Use \code{reference} in [radius_sum()] instead.}

\item{value}{Column name with values to sum.}

\item{lon_sub}{Deprecated. Use \code{lon_targets} instead.}

\item{lat_sub}{Deprecated. Use \code{lat_targets} instead.}

\item{lon_full}{Deprecated. Use \code{lon_reference} instead.}

\item{lat_full}{Deprecated. Use \code{lat_reference} instead.}

\item{radius}{Radius in meters.}

\item{display_progress}{Deprecated. Use \code{progress} in
\code{radius_sum()} instead.}

\item{result_col}{Name of the output column.}
}
\value{
See [radius_sum()].
}
\description{
`concentration()` is deprecated. Use [radius_sum()] instead.
}
\keyword{internal}
