% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration_cpp.R
\name{concentration_cpp}
\alias{concentration_cpp}
\title{Concentration cpp}
\usage{
concentration_cpp(sub, ref)
}
\arguments{
\item{sub}{deelverzameling}

\item{ref}{origineel}
}
\value{
df
}
\description{
Concentration cpp
}
