% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concentration_per_candidate_cell}
\alias{concentration_per_candidate_cell}
\title{Determine the concentrations within the highest focal cells for the current
iteration}
\usage{
concentration_per_candidate_cell(
  candidate_cells,
  dff,
  value,
  size,
  points,
  cache,
  radius,
  crs_metric,
  lon,
  lat
)
}
\arguments{
\item{candidate_cells}{data.frame containing cell ids with focal values
selected for refinement in the current iteration.}

\item{dff}{data.frame with all observations}

\item{value}{column name in `dff` to find concentrations for.}

\item{size}{size of cell in meters.}

\item{points}{number of points per `size`.}

\item{cache}{data.frame containing previously saved highest concentrations.}

\item{radius}{radius of circle in meters.}

\item{crs_metric}{metric CRS used for candidate cell coordinates.}

\item{lon}{longitude column name.}

\item{lat}{latitude column name.}
}
\description{
Determine the concentrations within the highest focal cells for
the current iteration.
}
\author{
Martin Haringa
}
\keyword{internal}
