% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration_risk.R
\name{concentration_risk}
\alias{concentration_risk}
\title{Title}
\usage{
concentration_risk(data, value, radius = 200)
}
\arguments{
\item{radius}{}
}
\description{
Title
}
\examples{
Groningen \%>\%
  filter(value > 1000) \%>\%
  concentration_risk(., value)

}
