% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_df_to_sf}
\alias{convert_df_to_sf}
\title{Convert data.frame to simple features (sf) object}
\usage{
convert_df_to_sf(df, lon = "lon", lat = "lat", crs_from = 4326, crs_to = 3035)
}
\arguments{
\item{df}{data.frame containing longitude and latitude columns}

\item{lon}{column name for longitude values (default: "lon").}

\item{lat}{column name for latitude values (default: "lat").}

\item{crs_from}{crs of the original coordinate system (default: 4326).}

\item{crs_to}{crs of the target coordinate system (default: 3035).}
}
\value{
Returns an sf object with the specified coordinate reference system.
}
\description{
This function converts a data.frame to a simple features (sf) object.
}
\author{
Martin Haringa
}
\keyword{internal}
