% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_in_circle.R
\name{coord_in_circle}
\alias{coord_in_circle}
\title{Coordinates in circle}
\usage{
coord_in_circle(data, lon = lon, lat = lat, radius = 200)
}
\arguments{
\item{lon}{Longitude of centre point, in degrees.}

\item{lat}{Latitude of centre point, in degrees.}

\item{radius}{Major (equatorial) radius (default is meters) of the ellipsoid. The default value is for WGS84.}
}
\value{
A data.frame of coordinates within \code{radius} around \code{(lon, lat)}.
}
\description{
Coordinates in circle
}
\examples{
#
coord_in_circle(Groningen, lon = 6.5203, lat = 53.24007, radius = 200)
}
