% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_gemeente.R
\docType{data}
\name{europe_countries}
\alias{europe_countries}
\title{Object of class \code{sf} for countries of Europe}
\format{
A simple feature object with 51 rows and 29 variables.
}
\usage{
europe_countries
}
\description{
An object of class \code{sf} (simple feature) for countries of Europe
}
\details{
The epsg (SRID) is set to 102013 (Europe Albers Equal Area Conic).
}
\author{
Martin Haringa
}
\keyword{datasets}
