% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{highest_concentration_candidate}
\alias{highest_concentration_candidate}
\title{Find the highest concentration for the current iteration}
\usage{
highest_concentration_candidate(new_candidates, candidate_cells, cache)
}
\arguments{
\item{new_candidates}{highest concentrations from the current iteration.}

\item{candidate_cells}{data.frame containing cell ids with focal values
selected for refinement in the current iteration.}

\item{cache}{data.frame containing previously saved highest concentrations.}
}
\description{
Find the highest concentration for the current iteration.
}
\author{
Martin Haringa
}
\keyword{internal}
