% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-aliases.R
\name{interpolate_spline}
\alias{interpolate_spline}
\title{Interpolate values using spherical splines}
\usage{
interpolate_spline(
  observations,
  targets,
  value,
  lon_obs = lon,
  lat_obs = lat,
  lon_targets = lon,
  lat_targets = lat,
  k = 50
)
}
\arguments{
\item{observations}{data.frame of observations.}

\item{targets}{data.frame of locations to calculate the interpolated and
smoothed values for.}

\item{value}{Column with values in \code{observations}.}

\item{lon_obs}{Column in \code{observations} with longitude.}

\item{lat_obs}{Column in \code{observations} with latitude.}

\item{lon_targets}{Column in \code{targets} with longitude.}

\item{lat_targets}{Column in \code{targets} with latitude.}

\item{k}{Basis dimension. For small data sets reduce \code{k} manually.}
}
\value{
Object equal to \code{targets} with an extra prediction column.
}
\description{
Deprecated. Spline interpolation and smoothing on the sphere.
This function is outside the main scope of \pkg{spatialrisk} and will be
removed in a future release.
}
\references{
\code{\link[mgcv:smooth.construct.sos.smooth.spec]{Splines on
the sphere}}
}
\keyword{internal}
