% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmi_historic_data.R
\name{knmi_historic_data}
\alias{knmi_historic_data}
\title{Retrieve historic weather data for the Netherlands}
\format{
The returned data frame contains the following columns:
\itemize{
  \item station = ID of measurement station;
  \item date = Date;
  \item FH = Hourly mean wind speed (in 0.1 m/s);
  \item FX = Maximum wind gust (in 0.1 m/s) during the hourly division;
  \item DR = Precipitation duration (in 0.1 hour) during the hourly division;
  \item RH = Hourly precipitation amount (in 0.1 mm) (-1 for <0.05 mm);
  \item city = City where the measurement station is located;
  \item lon = Longitude of station (EPSG:4326);
  \item lat = Latitude of station (EPSG:4326).
}
}
\usage{
knmi_historic_data(
  startyear,
  endyear,
  stations = NULL,
  progress = interactive()
)
}
\arguments{
\item{startyear, endyear}{Start and end year for the historic weather data.
Both must be single whole years. KNMI hourly data is available from 1951.}

\item{stations}{Optional station IDs to download. The default,
\code{NULL}, downloads all stations in \code{\link{knmi_stations}}.}

\item{progress}{Should a progress bar be shown? Defaults to
\code{interactive()}.}
}
\value{
Data frame containing weather data and metadata for weather station
locations.
}
\description{
This function retrieves historic hourly weather data collected by official
KNMI weather stations. See \code{\link{knmi_stations}} for the station
metadata included in this package.
}
\details{
The data is downloaded from KNMI when the function is called. This
requires an internet connection and may take some time when many stations or
years are requested.
}
\examples{
\dontrun{
knmi_historic_data(2015, 2019, stations = c(260, 280))
}

}
\author{
Martin Haringa
}
