% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lower_bound_fn}
\alias{lower_bound_fn}
\title{Determine a lower bound for highest concentration
Calculate concentration for the coordinates of the highest insured sums}
\usage{
lower_bound_fn(df, col, lat_nm, lon_nm, radius = 200, highest = 50)
}
\description{
Determine a lower bound for highest concentration
Calculate concentration for the coordinates of the highest insured sums
}
\keyword{internal}
