% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mw_create}
\alias{mw_create}
\title{Create focal ("moving window") weight matrix}
\usage{
mw_create(r, radius)
}
\arguments{
\item{r}{SpatRaster.}

\item{radius}{radius of the circle (in units of the crs).}
}
\description{
Create a focal ("moving window") weight matrix for use in
\code{terra::focal()}.
}
\details{
The returned mask includes raster-cell centres whose Euclidean
  distance from the focal-cell centre is no greater than `radius` plus the
  full raster-cell diagonal. This conservative expansion accounts for the
  possible displacement of both a disk centre and a contributing point from
  their respective cell centres.
}
\author{
Martin Haringa
}
\keyword{internal}
