% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-spatial-nl.R
\docType{data}
\name{nl_corop}
\alias{nl_corop}
\title{COROP regions in the Netherlands}
\format{
A simple feature object with 40 rows and 5 variables:
\describe{
  \item{corop_nr}{COROP number.}
  \item{areaname}{COROP region name.}
  \item{geometry}{COROP region geometry.}
  \item{lon}{Longitude of the COROP centroid.}
  \item{lat}{Latitude of the COROP centroid.}
}
}
\source{
Statistics Netherlands (CBS), adapted for package examples.
}
\usage{
nl_corop
}
\description{
An \code{sf} object with COROP region geometries for the
Netherlands. Centroid coordinates are included in EPSG:4326.
}
\details{
COROP regions are regional areas used for analytical purposes by,
among others, Statistics Netherlands.
}
\author{
Martin Haringa
}
\keyword{datasets}
