% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-spatial-nl.R
\docType{data}
\name{nl_gemeente}
\alias{nl_gemeente}
\title{Municipalities in the Netherlands}
\format{
A simple feature object with 380 rows and 6 variables:
\describe{
  \item{id}{Municipality identifier.}
  \item{code}{Municipality code.}
  \item{areaname}{Municipality name.}
  \item{lon}{Longitude of the municipality centroid.}
  \item{lat}{Latitude of the municipality centroid.}
  \item{geometry}{Municipality geometry.}
}
}
\source{
Statistics Netherlands (CBS), adapted for package examples.
}
\usage{
nl_gemeente
}
\description{
An \code{sf} object with municipal geometries for the
Netherlands in 2021. Centroid coordinates are included in EPSG:4326.
}
\author{
Martin Haringa
}
\keyword{datasets}
