% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-spatial-nl.R
\docType{data}
\name{nl_postcode2}
\alias{nl_postcode2}
\title{Two-digit postcode regions in the Netherlands}
\format{
A simple feature object with 90 rows and 4 variables:
\describe{
  \item{areaname}{2-digit postcode area.}
  \item{geometry}{Postcode region geometry.}
  \item{lon}{Longitude of the 2-digit postcode centroid.}
  \item{lat}{Latitude of the 2-digit postcode centroid.}
}
}
\source{
Adapted from Dutch postcode boundary data for package examples.
}
\usage{
nl_postcode2
}
\description{
An \code{sf} object with 2-digit postcode region geometries for
the Netherlands. Centroid coordinates are included in EPSG:4326.
}
\details{
Postal codes in the Netherlands are alphanumeric and consist of four
digits followed by two uppercase letters. This object aggregates those codes
to their first two digits.
}
\author{
Martin Haringa
}
\keyword{datasets}
