% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-spatial-nl.R
\docType{data}
\name{nl_postcode4}
\alias{nl_postcode4}
\title{Four-digit postcode regions in the Netherlands}
\format{
A simple feature object with 4053 rows and 7 variables:
\describe{
  \item{pc4}{4-digit postcode.}
  \item{areaname}{Name of corresponding 4-digit postcode area.}
  \item{city}{City name.}
  \item{biggest_20cities}{Whether the area is in one of the twenty largest
  cities in the Netherlands.}
  \item{geometry}{Postcode region geometry.}
  \item{lon}{Longitude of the 4-digit postcode centroid.}
  \item{lat}{Latitude of the 4-digit postcode centroid.}
}
}
\source{
Adapted from Dutch postcode boundary data for package examples.
}
\usage{
nl_postcode4
}
\description{
An \code{sf} object with 4-digit postcode region geometries for
the Netherlands. Centroid coordinates are included in EPSG:4326.
}
\details{
Postal codes in the Netherlands are alphanumeric and consist of four
digits followed by two uppercase letters. This object aggregates those codes
to their first four digits.
}
\author{
Martin Haringa
}
\keyword{datasets}
