% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-spatial-nl.R
\docType{data}
\name{nl_provincie}
\alias{nl_provincie}
\title{Provinces in the Netherlands}
\format{
A simple feature object with 12 rows and 4 variables:
\describe{
  \item{areaname}{Province name.}
  \item{geometry}{Province geometry.}
  \item{lon}{Longitude of the province centroid.}
  \item{lat}{Latitude of the province centroid.}
}
}
\source{
Statistics Netherlands (CBS), adapted for package examples.
}
\usage{
nl_provincie
}
\description{
An \code{sf} object with province geometries for the Netherlands.
Centroid coordinates are included in EPSG:4326.
}
\author{
Martin Haringa
}
\keyword{datasets}
