% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_concentration.R
\name{plot.neighborhood}
\alias{plot.neighborhood}
\title{Automatically create a plot for objects obtained from neighborhood_gh_search()}
\usage{
\method{plot}{neighborhood}(x, buffer = 0, legend_title = NULL, ...)
}
\arguments{
\item{x}{object neighborhood object produced by `neighborhood_gh_search()`}

\item{buffer}{numeric value, show objects within buffer (in meters) from circle (defaults to 0)}

\item{legend_title}{title of legend}

\item{...}{additional arguments affecting the interactive map produced}
}
\value{
Interactive view of highest concentration on map
}
\description{
Takes an object produced by `neighborhood_gh_search()`, and creates an interactive map.
}
\author{
Martin Haringa
}
