% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-aliases.R
\name{points_in_circle}
\alias{points_in_circle}
\title{Deprecated alias for points_within_radius()}
\usage{
points_in_circle(
  data,
  lon_center,
  lat_center,
  lon = lon,
  lat = lat,
  radius = 200,
  sort = TRUE
)
}
\arguments{
\item{data}{A data.frame containing point coordinates.}

\item{lon_center}{Longitude of the circle centre.}

\item{lat_center}{Latitude of the circle centre.}

\item{lon}{Column name with longitude.}

\item{lat}{Column name with latitude.}

\item{radius}{Radius in meters.}

\item{sort}{Logical. Whether to sort output by distance.}
}
\value{
See [points_within_radius()].
}
\description{
`points_in_circle()` is deprecated. Use [points_within_radius()] instead.
}
\keyword{internal}
