% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-aliases.R
\name{points_to_polygon}
\alias{points_to_polygon}
\title{Deprecated alias for summarise_points_by_polygon()}
\usage{
points_to_polygon(sf_map, df, oper, crs = 4326, outside_print = FALSE)
}
\arguments{
\item{sf_map}{Deprecated. Use \code{polygons} instead.}

\item{df}{Deprecated. Use \code{points} instead.}

\item{oper}{Deprecated expression used to aggregate values.}

\item{crs}{Coordinate reference system of the point coordinates.}

\item{outside_print}{Deprecated. Use \code{outside} instead.}
}
\value{
See [summarise_points_by_polygon()].
}
\description{
`points_to_polygon()` is deprecated. Use [summarise_points_by_polygon()]
instead.
}
\keyword{internal}
