% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_concentration_terra.R
\name{print.concentration}
\alias{print.concentration}
\title{Print output}
\usage{
\method{print}{concentration}(x, ...)
}
\arguments{
\item{x}{input from function}

\item{...}{additional arguments}
}
\description{
Print output
}
