% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_sphere.R
\name{smooth_sphere}
\alias{smooth_sphere}
\title{Splines on the sphere}
\usage{
smooth_sphere(df, sf_object, value, lon_df = lon, lat_df = lat,
  lon_sf = lon, lat_sf = lat)
}
\arguments{
\item{df}{data.frame of observations.}

\item{sf_object}{Object of locations to calculate the smoothed values for (target points).}

\item{value}{Column with values in \code{df}.}

\item{lon_df}{Column in \code{df} with longitude (lon is default).}

\item{lat_df}{Column in \code{df} with latitude (lat is default).}

\item{lon_sf}{Column in \code{sf} with longitude (lon is default).}

\item{lat_sf}{Column in \code{sf} with latitude (lat is default).}
}
\value{
Object equal to object \code{sf_object} including an extra column \code{prediction}.
}
\description{
Spline interpolation and smoothing on the sphere.
}
\details{
\code{df} should include at least columns for longitude and latitude.

\code{sf_object} should include at least columns for longitude, latitude and value of interest to interpolate and smooth.

A smooth of the general type discussed in Duchon (1977) is used: the sphere is embedded in a 3D Euclidean space, but smoothing employs a penalty based on second derivatives (so that locally as the smoothing parameter tends to zero we recover a "normal" thin plate spline on the tangent space). This is an unpublished suggestion of Jean Duchon.
}
\examples{
\dontrun{
pop_sf <- smooth_sphere(insurance, nl_postcode3, population_pc4)
choropleth(pop_sf, value = "prediction", n = 13)
}

}
\references{
\code{\link[mgcv:smooth.construct.sos.smooth.spec]{Splines on the sphere}}
}
\author{
Martin Haringa
}
