% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-aliases.R
\name{spatialrisk-deprecated}
\alias{spatialrisk-deprecated}
\alias{find_highest_concentration}
\title{Deprecated aliases}
\usage{
find_highest_concentration(
  df,
  value,
  top_n = 1,
  radius = 200,
  cell_size = 100,
  grid_precision = 1,
  lon = "lon",
  lat = "lat",
  crs_metric = 3035,
  print_progress = TRUE
)
}
\arguments{
\item{df}{Deprecated wrapper argument.}

\item{value}{Deprecated wrapper argument.}

\item{top_n}{Deprecated wrapper argument.}

\item{radius}{Deprecated wrapper argument.}

\item{cell_size}{Deprecated wrapper argument.}

\item{grid_precision}{Deprecated wrapper argument.}

\item{lon}{Deprecated wrapper argument.}

\item{lat}{Deprecated wrapper argument.}

\item{crs_metric}{Deprecated wrapper argument.}

\item{print_progress}{Deprecated wrapper argument.

`find_highest_concentration()` was renamed to [concentration_hotspot()].}
}
\description{
These functions are deprecated compatibility wrappers. Use the replacement
functions shown in the warning messages.

Deprecated.
}
