% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_in_circle.R
\name{sum_in_circle}
\alias{sum_in_circle}
\title{Title}
\usage{
sum_in_circle(data, value, lon, lat, radius = 200)
}
\arguments{
\item{data}{Data frame}

\item{value}{waarde}

\item{lon}{longitude}

\item{lat}{latitude}

\item{radius}{radius in meters}
}
\value{
waarde
}
\description{
Title
}
\examples{
sum_in_circle(Groningen, value, 6.520386, 53.24007)
}
