% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{top_n_focals}
\alias{top_n_focals}
\title{Identify the focal indices with the highest values}
\usage{
top_n_focals(focal, n)
}
\arguments{
\item{focal}{focal as output from \code{terra::focal()}.}

\item{n}{positive integer value greater or equal to 1.}
}
\description{
Generate a data.frame containing the cell indices of the focal
cells with the \code{n} highest values. Additionally, include columns for
the coordinates \code{(xy)} corresponding to the center of each cell.
}
\author{
Martin Haringa
}
\keyword{internal}
