% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_db}
\alias{update_db}
\title{Save highest concentrations per cell for subsequent iterations}
\usage{
update_db(hf_conc_new, db, cells)
}
\arguments{
\item{hf_conc_new}{highest concentrations from the current iteration,
obtained from \code{conc_per_cell_new()}.}

\item{db}{data.frame containing previously saved highest concentrations.}

\item{cells}{cells containing points associated with the current highest
concentration to be removed from \code{db}.}
}
\description{
Save highest concentrations per cell for subsequent iterations.
}
\author{
Martin Haringa
}
\keyword{internal}
