% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_hotspot_cache}
\alias{update_hotspot_cache}
\title{Save highest concentrations per cell for subsequent iterations}
\usage{
update_hotspot_cache(new_candidates, cache, cells)
}
\arguments{
\item{new_candidates}{highest concentrations from the current iteration.}

\item{cache}{data.frame containing previously saved highest concentrations.}

\item{cells}{cells containing points associated with the current highest
concentration to be removed from \code{cache}.}
}
\description{
Save highest concentrations per cell for subsequent iterations.
}
\author{
Martin Haringa
}
\keyword{internal}
