% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_gemeente.R
\docType{data}
\name{world_countries}
\alias{world_countries}
\title{Object of class sf for countries of the entire world}
\format{A simple feature object with 234 rows and 29 variables.}
\usage{
world_countries
}
\description{
An object of class sf (simple feature) for countries of the entire world.
}
\author{
Martin Haringa
}
\keyword{datasets}
