% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_around_coord.R
\name{block_around_coord}
\alias{block_around_coord}
\title{Coordinates of the four cardinal directions from centre}
\usage{
block_around_coord(lon = lon, lat = lat, radius = 200)
}
\arguments{
\item{lon}{Longitude of point(s), in degrees.}

\item{lat}{Latitude of point(s), in degrees.}

\item{radius}{Major (equatorial) radius (default is meters) of the ellipsoid. The default value is for WGS84.}
}
\value{
A data.frame consisting of the four coordinates (long,lat) obtained by walking the radius to the four cardinal directions.
}
\description{
Coordinates of the four cardinal directions from centre
}
\examples{
block_around_coord(lon = 6.520386, lat = 53.24007, radius = 200)
}
