% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-aliases.R
\name{choropleth_ggplot2}
\alias{choropleth_ggplot2}
\title{Choropleth map of an sf object with ggplot2}
\usage{
choropleth_ggplot2(
  sf_object,
  value = output,
  n = 7,
  dig.lab = 2,
  legend_title = "Class",
  option = "D",
  direction = 1
)
}
\arguments{
\item{sf_object}{An object of class \code{sf} containing polygon geometries.}

\item{value}{Column in \code{sf_object} used to shade the polygons.}

\item{n}{Integer. Number of clusters to use in Fisher classification.}

\item{dig.lab}{Integer. Number of digits to display in legend labels.}

\item{legend_title}{Character. Title for the legend.}

\item{option}{Character string indicating the colormap option passed to
\code{ggplot2}.}

\item{direction}{Numeric. Order of colors in the scale.}
}
\value{
A \code{ggplot} object containing the choropleth map.
}
\description{
Deprecated. Use [choropleth()] instead.
}
\details{
`choropleth_ggplot2()` is deprecated. Use [choropleth()] instead.
}
