% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ggplot2.R
\name{choropleth_ggplot2}
\alias{choropleth_ggplot2}
\title{Map object of class sf using ggplot2}
\usage{
choropleth_ggplot2(sf_object, value = output, n = 7, dig.lab = 2,
  legend_title = "Class")
}
\arguments{
\item{sf_object}{object of class sf}

\item{value}{column to shade the polygons}

\item{n}{number of clusters (default is 7)}

\item{dig.lab}{number of digits in legend (default is 2)}

\item{legend_title}{title of legend}
}
\value{
ggplot map
}
\description{
Takes an object produced by \code{choropleth_sf()}, and creates the correspoding choropleth map.
}
\examples{
test <- choropleth_sf(nl_postcode2, insurance, sum(amount, na.rm = TRUE))
choropleth_ggplot2(test)
}
\author{
Martin Haringa
}
