% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration.R
\name{concentration}
\alias{concentration}
\title{Concentration}
\usage{
concentration(data, value, radius = 200, lon = lon, lat = lat)
}
\arguments{
\item{data}{Data frame}

\item{value}{Column name with value}

\item{radius}{Radius (in meters) (default is 200m)}

\item{lon}{Column name with longitude (lon is default)}

\item{lat}{Column name with latitude (lat is default)}
}
\value{
value
}
\description{
Concentration
}
\examples{
concentration(Groningen[1:10, ], amount, radius = 1000, lon = lon, lat = lat)

}
