% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration.R
\name{concentration}
\alias{concentration}
\title{Concentration risk}
\usage{
concentration(sub, full, lon_sub = lon, lat_sub = lat,
  value_sub = value, lon_full = lon, lat_full = lat,
  value_full = value, radius = 200, display_progress = TRUE)
}
\arguments{
\item{sub}{data.frame of locations to calculate concentration risk for. The data.frame should include at least columns for longitude, latitude and value of interest to summarize.}

\item{full}{data.frame of full portfolio. The data.frame should include at least columns for longitude, latitude and value of interest to summarize.}

\item{lon_sub}{Column name with longitude (lon is default)}

\item{lat_sub}{Column name with latitude (lat is default)}

\item{value_sub}{Column name with value in the sub data (value is default)}

\item{lon_full}{Column name with longitude in the full data (lon is default)}

\item{lat_full}{Column name with latitude in the full data (lat is default)}

\item{value_full}{Column name with value in the full data (value is default)}

\item{radius}{Radius (in meters) (default is 200m)}

\item{display_progress}{Show progress bar (TRUE/FALSE)}
}
\value{
value
}
\description{
Concentration risk
}
\examples{
concentration(Groningen[1:10,], Groningen, value_sub = amount, value_full = amount)

}
\author{
Martin Haringa
}
