% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration.R
\name{concentration}
\alias{concentration}
\title{Concentration}
\usage{
concentration(data, value, radius = 200)
}
\arguments{
\item{data}{Data frame}

\item{value}{Column name}

\item{radius}{Radius (in meters)}
}
\value{
value
}
\description{
Concentration
}
\examples{
concentration(Groningen[1:10, ], amount, radius = 1000)

}
