% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration.R
\name{concentration}
\alias{concentration}
\title{Concentration risk}
\usage{
concentration(sub, full, lon_sub = lon, lat_sub = lat,
  lon_full = lon, lat_full = lat, value_full = value, radius = 200,
  display_progress = TRUE)
}
\arguments{
\item{sub}{data.frame of locations to calculate concentration risk for. The data.frame should include at least columns for longitude, latitude and value of interest to summarize.}

\item{full}{data.frame of full portfolio. The data.frame should include at least columns for longitude, latitude and value of interest to summarize.}

\item{lon_sub}{Column name with longitude (lon is default).}

\item{lat_sub}{Column name with latitude (lat is default).}

\item{lon_full}{Column name with longitude in the full data (lon is default).}

\item{lat_full}{Column name with latitude in the full data (lat is default).}

\item{value_full}{Column name with value in the full data (value is default).}

\item{radius}{Radius (in meters) (default is 200m).}

\item{display_progress}{Show progress bar (TRUE/FALSE).}
}
\value{
A data.frame equal to data.frame \code{sub} including an extra column \code{concentration}.
}
\description{
The points (e.g. addresses) with the highest concentrations. In the context of the 'standard formula' under Solvency II,
this is the maximum gross sum insured of the set of buildings fully or partly located within this radius.
}
\examples{
df <- data.frame(location = c("p1", "p2"), lon = c(6.561561, 6.561398), lat = c(53.21369, 53.21326))
concentration(df, Groningen, value = amount, radius = 100)

}
\author{
Martin Haringa
}
