% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration_risk.R
\name{concentration_risk}
\alias{concentration_risk}
\title{Concentration risk}
\usage{
concentration_risk(data, value, radius = 200)
}
\arguments{
\item{data}{Data frame}

\item{value}{Column}

\item{radius}{Radius (in meters)}
}
\value{
value
}
\description{
Concentration risk
}
\examples{
Groningen \%>\%
  filter(value > 1000) \%>\%
  concentration_risk(., value)

}
