% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_in_circle.R
\name{coord_in_circle}
\alias{coord_in_circle}
\title{Coordinates in circle}
\usage{
coord_in_circle(data, lon = lon, lat = lat, radius = 200)
}
\arguments{
\item{radius}{}
}
\description{
Coordinates in circle
}
