% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_concentration_terra.R
\name{find_highest_concentration}
\alias{find_highest_concentration}
\title{Find highest concentration}
\usage{
find_highest_concentration(
  df,
  value,
  top_n = 1,
  radius = 200,
  cell_size = 100,
  grid_precision = 1
)
}
\arguments{
\item{df}{data.frame. Should include at least columns for longitude,
latitude, and the value of interest to summarize.}

\item{value}{column name with value of interest to summarize in \code{df}.}

\item{top_n}{positive integer value greater or equal to 1 (default is 1).}

\item{radius}{numeric. Radius of the circle in meters (default is 200).}

\item{cell_size}{numeric. Size of cell in meters (default is 100).}

\item{grid_precision}{numeric. Precision of grid in meters (default is 1).}
}
\value{
list
}
\description{
Determines the central coordinates of a circle with a constant
radius that maximizes the coverage of insured fire risk, where the fire risk
is defined as assuming 100 percent damage on the total sum of the capital
insured for each building located partly or fully within this radius.

A recent regulation by the European Commission mandates insurance companies
to report the maximum value of insured fire risk policies for all buildings
partially or fully situated within a circle with a radius of 200 meters.
}
\details{
This problem resembles a Maximal Covering Location Problem (MCLP)
with a fixed radius, belonging to the category of facility location problems.
The main aim is to select the best locations for a predetermined number of
facilities to achieve maximum coverage of demand points within a specified
radius of each facility. In essence, the objective is to identify optimal
facility locations to cover as many demand points as feasible, while ensuring
that each demand point falls within the designated distance (radius) of at
least one facility.
}
\examples{
x <- find_highest_concentration(Groningen, "amount")
plot(x)


}
\author{
Martin Haringa
}
