% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haversine.R
\name{haversine}
\alias{haversine}
\title{Haversine}
\usage{
haversine(lat_from, lon_from, lat_to, lon_to)
}
\arguments{
\item{lat_from}{Origin latitude}

\item{lon_from}{Origin longitude}

\item{lat_to}{Destination latitude}

\item{lon_to}{Destination longitude}
}
\value{
A data frame with new location
}
\description{
Haversine
}
\examples{
haversine(53.24007, 6.520386, 53.24054, 6.520386)

}
