% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_gemeente.R
\docType{data}
\name{nl_corop}
\alias{nl_corop}
\title{Object of class \code{sf} for COROP regions in the Netherlands}
\format{
A simple feature object with 40 rows and 5 variables:
\describe{
  \item{corop_nr}{corop number}
  \item{areaname}{corop name}
  \item{geometry}{geometry object of COROP region}
  \item{lon}{longitude of the corop centroid}
  \item{lat}{latitude of the corop centroid}
}
}
\usage{
nl_corop
}
\description{
An object of class \code{sf} (simple feature) for COROP regions
in the Netherlands.
}
\details{
A COROP region is a regional area within the Netherlands. These
regions are used for analytical purposes by, among others, Statistics
Netherlands. The Dutch abbreviation stands for Coordinatiecommissie Regionaal
Onderzoeksprogramma, literally the Coordination Commission Regional Research
Programme.
}
\author{
Martin Haringa
}
\keyword{datasets}
