% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_gemeente.R
\docType{data}
\name{nl_gemeente}
\alias{nl_gemeente}
\title{Object of class \code{sf} for municipalities in the Netherlands}
\format{
A simple feature object with 380 rows and 6 variables:
\describe{
  \item{id}{id of gemeente}
  \item{code}{code of gemeente}
  \item{areaname}{name of gemeente}
  \item{lon}{longitude of the gemeente centroid}
  \item{lat}{latitude of the gemeente centroid}
  \item{geometry}{geometry object of gemeente}
}
}
\usage{
nl_gemeente
}
\description{
An object of class \code{sf} (simple feature) for municipalities (Dutch: gemeentes) in the Netherlands in the year 2021.
}
\author{
Martin Haringa
}
\keyword{datasets}
