% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_gemeente.R
\docType{data}
\name{nl_gemeente}
\alias{nl_gemeente}
\title{Object of class sf for gemeentes (municipalities) in the Netherlands}
\format{A simple feature object with 380 rows and 4 variables:
\describe{
  \item{id}{id of gemeente}
  \item{code}{code of gemeente}
  \item{areaname}{name of gemeente}
  \item{geometry}{geometry object of gemeente}
}}
\usage{
nl_gemeente
}
\description{
An object of class sf (simple feature) for gemeentes (English: municipalities) in the Netherlands in the year 2018.
}
\author{
Martin Haringa
}
\keyword{datasets}
