% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_gemeente.R
\docType{data}
\name{nl_postcode2}
\alias{nl_postcode2}
\title{Object of class sf for 2-digit postcode regions in the Netherlands}
\format{A simple feature object with 90 rows and 2 variables:
\describe{
  \item{pc2}{2-digit postal code}
  \item{geometry}{geometry object of postal code}
}}
\usage{
nl_postcode2
}
\description{
An object of class sf (simple feature) for 2-digit postcode (English: postal code) regions in the Netherlands.
}
\details{
Postal codes in the Netherlands, known as postcodes, are alphanumeric, consisting of four digits followed by two uppercase letters. The first two digits indicate a city and a region, the second two digits and the two letters indicate a range of house numbers, usually on the same street.
}
\author{
Martin Haringa
}
\keyword{datasets}
