% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_gemeente.R
\docType{data}
\name{nl_postcode3}
\alias{nl_postcode3}
\title{Object of class \code{sf} for 3-digit postcode regions in the Netherlands}
\format{
A simple feature object with 799 rows and 3 variables:
\describe{
  \item{areaname}{3-digit postal code}
  \item{geometry}{geometry object of postal code}
  \item{lon}{longitude of the 3-digit postal code centroid}
  \item{lat}{latitude of the 3-digit postal code centroid}
}
}
\usage{
nl_postcode3
}
\description{
An object of class \code{sf} (simple feature) for 3-digit postal codes (Dutch: postcode) regions in the Netherlands.
}
\details{
Postal codes in the Netherlands, known as postcodes, are alphanumeric, consisting of four digits followed by two uppercase letters. The first two digits indicate a city and a region, the second two digits and the two letters indicate a range of house numbers, usually on the same street.
}
\author{
Martin Haringa
}
\keyword{datasets}
