% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_gemeente.R
\docType{data}
\name{nl_provincie}
\alias{nl_provincie}
\title{Object of class sf for provincies (provinces) in the Netherlands}
\format{A simple feature object with 12 rows and 2 variables:
\describe{
  \item{areaname}{province name}
  \item{geometry}{geometry object of province}
}}
\usage{
nl_provincie
}
\description{
An object of class sf (simple feature) for provincies (English: provinces) in the Netherlands.
}
\author{
Martin Haringa
}
\keyword{datasets}
