% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nl_gemeente.R
\docType{data}
\name{nl_provincie}
\alias{nl_provincie}
\title{Object of class \code{sf} for provinces in the Netherlands}
\format{
A simple feature object with 12 rows and 4 variables:
\describe{
  \item{areaname}{province name}
  \item{geometry}{geometry object of province}
  \item{lon}{longitude of the province centroid}
  \item{lat}{latitude of the province centroid}
}
}
\usage{
nl_provincie
}
\description{
An object of class \code{sf} (simple feature) for provinces
(Dutch: provincies) in the Netherlands.
}
\author{
Martin Haringa
}
\keyword{datasets}
