% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_highest_concentration.R
\name{plot.conc}
\alias{plot.conc}
\title{Automatically create a plot for objects obtained from highest_concentration()}
\usage{
\method{plot}{conc}(
  x,
  grid_points = TRUE,
  legend_title = NULL,
  palette = "viridis",
  legend_position = "bottomleft",
  providers = c("CartoDB.Positron", "nlmaps.luchtfoto"),
  ...
)
}
\arguments{
\item{x}{object of class `conc` obtained from
`highest_concentration()`}

\item{grid_points}{show grid points (TRUE), or objects (FALSE)}

\item{legend_title}{title of legend}

\item{palette}{palette for grid points (defaults to "viridis")}

\item{legend_position}{legend position for grid points legend (defaults to
"bottomleft")}

\item{providers}{providers to show. See `leaflet::providers` for a list.}

\item{...}{additional arguments affecting the interactive map produced}
}
\value{
Interactive view of geohashes with highest concentrations
}
\description{
Takes an object produced by `highest_concentration()`,
and creates an interactive map.
}
\author{
Martin Haringa
}
