% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot_points.R
\name{plot_points}
\alias{plot_points}
\title{Create map with points}
\usage{
plot_points(df, value, lon = "lon", lat = "lat", crs = 4326, at = NULL)
}
\arguments{
\item{df}{data.frame containing columns for longitude and latitude.}

\item{value}{column in \code{df} to be visualized.}

\item{lon}{column in \code{df} containing longitude values.}

\item{lat}{column in \code{df} containing latitude values.}

\item{crs}{crs code for the coordinate reference system (default is 4326).}

\item{at}{the breakpoints used for visualisation.}
}
\description{
Create map for a data.frame containing points.
}
\examples{
\dontrun{
plot_points(Groningen, value = "amount")
}

}
