% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{Create map with points}
\usage{
plot_points(
  df,
  value,
  lon = lon,
  lat = lat,
  palette = "viridis",
  legend_position = "bottomleft",
  crs = 4326
)
}
\arguments{
\item{df}{data.framw with column for lon and lat}

\item{value}{column in df}

\item{lon}{column with lon}

\item{lat}{column with lat}

\item{palette}{color palette}

\item{legend_position}{position for legend (default is "bottomleft")}

\item{crs}{crs (default is 4326)}
}
\value{
leaflet map
}
\description{
Create map for data.frame with points.
}
\examples{
\dontrun{
plot_points(Groningen, value = amount)
}

}
