% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_in_circle.R
\name{points_in_circle}
\alias{points_in_circle}
\title{Coordinates in circle}
\usage{
points_in_circle(data, lon_center, lat_center, lon = lon, lat = lat,
  value = value, radius = 200)
}
\arguments{
\item{data}{Data}

\item{lon_center}{Longitude of center point, in degrees.}

\item{lat_center}{Latitude of center point, in degrees.}

\item{lon}{Name of column in data with longitudes. The default name is lon.}

\item{lat}{Name of column in data with latitudes. The default name is lat.}

\item{value}{Name of column in data with insured sums.}

\item{radius}{Major (equatorial) radius (default is meters) of the ellipsoid. The default value is for WGS84.}
}
\value{
A data.frame of coordinates within \code{radius} around \code{(lon_center, lat_center)}.
}
\description{
Coordinates in circle
}
