% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_within_radius.R
\name{points_within_radius}
\alias{points_within_radius}
\title{Find points within radius around one or more center coordinates}
\usage{
points_within_radius(
  data,
  lon_center,
  lat_center,
  lon = "lon",
  lat = "lat",
  radius = 200,
  sort = TRUE
)
}
\arguments{
\item{data}{A data frame containing at least longitude and latitude columns.}

\item{lon_center}{Numeric scalar or vector, longitude(s) of the circle center(s).}

\item{lat_center}{Numeric scalar or vector, latitude(s) of the circle center(s).}

\item{lon}{A string with the name of the longitude column in `data`.}

\item{lat}{A string with the name of the latitude column in `data`.}

\item{radius}{Numeric, circle radius in meters. Default is 200.}

\item{sort}{Logical, if `TRUE` results are sorted by distance within each center.}
}
\value{
A data frame subset of `data` with an extra column `distance_m` and
  if multiple centers are provided, also a column `center_index`.
}
\description{
This function selects rows from a data frame whose
longitude/latitude coordinates fall within a given radius (in meters)
from one or more specified center points. It also calculates the distance
of each point to the center.
}
